requirements.txt
folium==0.16.0
fpdf2==2.7.9
kaleido==0.2.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
plotly==5.20.0
streamlit==1.35.0
XlsxWriter==3.2.0

